More Capturable methods into macro
