+++ aderyn_core/src/ast/macros.rs
@@ -97,7 +97,71 @@ macro_rules! node_group {
     };
 }
 
+macro_rules! generate_ast_methods {
+    ($( $name:ident ),* $(,)*) => {
+
+        #[derive(Debug, Clone, PartialEq)]
+        pub enum ASTNode {
+            $($name($name),)*
+        }
+
+        $(
+            impl From<$name> for ASTNode {
+                fn from(value: $name) -> Self {
+                    ASTNode::$name(value)
+                }
+            }
+
+            impl From<&$name> for ASTNode {
+                fn from(value: &$name) -> Self {
+                    ASTNode::$name(value.clone())
+                }
+            }
+        )*
+
+        impl ASTNode {
+            pub fn node_type(&self) -> NodeType {
+                match self {
+                    $(ASTNode::$name(_) => NodeType::$name,)*
+                }
+            }
+            pub fn id(&self) -> Option<NodeID> {
+                match self {
+                    $(ASTNode::$name(n) => Some(n.id),)*
+                }
+            }
+        }
+
+        impl Node for ASTNode {
+            fn accept(&self, visitor: &mut impl ASTConstVisitor) -> eyre::Result<()> {
+                match self {
+                    $(ASTNode::$name(n) => n.accept(visitor),)*
+                }
+            }
+            fn accept_metadata(&self, visitor: &mut impl ASTConstVisitor) -> eyre::Result<()> {
+                match self {
+                    $(ASTNode::$name(n) => n.accept_metadata(visitor),)*
+                }
+            }
+            fn accept_id(&self, visitor: &mut impl ASTConstVisitor) -> Result<()> {
+                visitor.visit_node_id(self.id())?;
+                Ok(())
+            }
+        }
+
+        impl ASTNode {
+            pub fn src(&self) -> Option<&str> {
+                match self {
+                    $(ASTNode::$name(node) => Some(&node.src),)*
+                }
+            }
+        }
+
+    };
+}
+
 pub(crate) use ast_node;
 pub(crate) use ast_node_no_partial_eq;
 pub(crate) use expr_node;
+pub(crate) use generate_ast_methods;
 pub(crate) use node_group;
