+++ aderyn_core/src/context/capturable.rs
@@ -1,59 +1,57 @@
 use crate::ast::{ContractDefinition, *};
 
 use super::{
-    macros::generate_capture_methods,
+    macros::generate_capturable_methods,
     workspace_context::{ASTNode, WorkspaceContext},
 };
 
-#[derive(Clone)]
-pub enum Capturable {
-    ASTNode(ASTNode),
-    Assignment(Assignment),
-    BinaryOperation(BinaryOperation),
-    Block(Block),
-    Conditional(Conditional),
-    ContractDefinition(ContractDefinition),
-    ASTContractDefinition(ContractDefinition),
-    ElementaryTypeNameExpression(ElementaryTypeNameExpression),
-    EnumDefinition(EnumDefinition),
-    EnumValue(EnumValue),
-    EventDefinition(EventDefinition),
-    ErrorDefinition(ErrorDefinition),
-    FunctionCall(FunctionCall),
-    FunctionCallOptions(FunctionCallOptions),
-    FunctionDefinition(FunctionDefinition),
-    ForStatement(ForStatement),
-    Identifier(Identifier),
-    IdentifierPath(IdentifierPath),
-    IfStatement(IfStatement),
-    ImportDirective(ImportDirective),
-    IndexAccess(IndexAccess),
-    IndexRangeAccess(IndexRangeAccess),
-    InheritanceSpecifier(InheritanceSpecifier),
-    InlineAssembly(InlineAssembly),
-    Literal(Literal),
-    MemberAccess(MemberAccess),
-    NewExpression(NewExpression),
-    ModifierDefinition(ModifierDefinition),
-    ModifierInvocation(ModifierInvocation),
-    OverrideSpecifier(OverrideSpecifier),
-    ParameterList(ParameterList),
-    PragmaDirective(PragmaDirective),
-    Return(Return),
-    SourceUnit(SourceUnit),
-    StructDefinition(StructDefinition),
-    StructuredDocumentation(StructuredDocumentation),
-    TupleExpression(TupleExpression),
-    UnaryOperation(UnaryOperation),
-    UserDefinedValueTypeDefinition(UserDefinedValueTypeDefinition),
-    UsingForDirective(UsingForDirective),
-    VariableDeclaration(VariableDeclaration),
-    VariableDeclarationStatement(VariableDeclarationStatement),
-    WhileStatement(WhileStatement),
-    DoWhileStatement(DoWhileStatement),
-    Break(Break),
-    Continue(Continue),
-    PlaceholderStatement(PlaceholderStatement),
+generate_capturable_methods! {
+    ASTNode,
+    Assignment,
+    BinaryOperation,
+    Block,
+    Conditional,
+    ContractDefinition,
+    ElementaryTypeNameExpression,
+    EnumDefinition,
+    EnumValue,
+    EventDefinition,
+    ErrorDefinition,
+    FunctionCall,
+    FunctionCallOptions,
+    FunctionDefinition,
+    ForStatement,
+    Identifier,
+    IdentifierPath,
+    IfStatement,
+    ImportDirective,
+    IndexAccess,
+    IndexRangeAccess,
+    InheritanceSpecifier,
+    InlineAssembly,
+    Literal,
+    MemberAccess,
+    NewExpression,
+    ModifierDefinition,
+    ModifierInvocation,
+    OverrideSpecifier,
+    ParameterList,
+    PragmaDirective,
+    Return,
+    SourceUnit,
+    StructDefinition,
+    StructuredDocumentation,
+    TupleExpression,
+    UnaryOperation,
+    UserDefinedValueTypeDefinition,
+    UsingForDirective,
+    VariableDeclaration,
+    VariableDeclarationStatement,
+    WhileStatement,
+    DoWhileStatement,
+    Break,
+    Continue,
+    PlaceholderStatement,
 }
 
 impl Capturable {
@@ -65,7 +63,6 @@ impl Capturable {
             Self::Block(n) => context.get_node_sort_key(&n.into()),
             Self::Conditional(n) => context.get_node_sort_key(&n.into()),
             Self::ContractDefinition(n) => context.get_node_sort_key(&n.into()),
-            Self::ASTContractDefinition(n) => context.get_node_sort_key(&n.into()),
             Self::ElementaryTypeNameExpression(n) => context.get_node_sort_key(&n.into()),
             Self::EnumDefinition(n) => context.get_node_sort_key(&n.into()),
             Self::EnumValue(n) => context.get_node_sort_key(&n.into()),
@@ -117,7 +114,6 @@ impl Capturable {
             Self::Block(n) => Some(n.id),
             Self::Conditional(n) => Some(n.id),
             Self::ContractDefinition(n) => Some(n.id),
-            Self::ASTContractDefinition(n) => Some(n.id),
             Self::ElementaryTypeNameExpression(n) => Some(n.id),
             Self::EnumDefinition(n) => Some(n.id),
             Self::EnumValue(n) => Some(n.id),
@@ -160,55 +156,6 @@ impl Capturable {
             Self::PlaceholderStatement(n) => Some(n.id),
         }
     }
-}
-
-generate_capture_methods! {
-    ASTNode,
-    Assignment,
-    BinaryOperation,
-    Block,
-    Conditional,
-    ContractDefinition,
-    ElementaryTypeNameExpression,
-    EnumDefinition,
-    EnumValue,
-    EventDefinition,
-    ErrorDefinition,
-    FunctionCall,
-    FunctionCallOptions,
-    FunctionDefinition,
-    ForStatement,
-    Identifier,
-    IdentifierPath,
-    IfStatement,
-    ImportDirective,
-    IndexAccess,
-    IndexRangeAccess,
-    InheritanceSpecifier,
-    InlineAssembly,
-    Literal,
-    MemberAccess,
-    NewExpression,
-    ModifierDefinition,
-    ModifierInvocation,
-    OverrideSpecifier,
-    ParameterList,
-    PragmaDirective,
-    Return,
-    SourceUnit,
-    StructDefinition,
-    StructuredDocumentation,
-    TupleExpression,
-    UnaryOperation,
-    UserDefinedValueTypeDefinition,
-    UsingForDirective,
-    VariableDeclaration,
-    VariableDeclarationStatement,
-    WhileStatement,
-    DoWhileStatement,
-    Break,
-    Continue,
-    PlaceholderStatement,
 }
 
 impl From<&&ContractDefinition> for Capturable {
