+++ aderyn_core/src/context/macros.rs
@@ -1,5 +1,11 @@
-macro_rules! generate_capture_methods {
+macro_rules! generate_capturable_methods {
     ($( $name:ident ),* $(,)*) => {
+
+        #[derive(Clone)]
+        pub enum Capturable {
+            $($name($name),)*
+        }
+
         $(
             impl From<$name> for Capturable {
                 fn from(value: $name) -> Self {
@@ -16,68 +22,4 @@ macro_rules! generate_capture_methods {
     };
 }
 
-macro_rules! generate_ast_methods {
-    ($( $name:ident ),* $(,)*) => {
-
-        #[derive(Debug, Clone, PartialEq)]
-        pub enum ASTNode {
-            $($name($name),)*
-        }
-
-        $(
-            impl From<$name> for ASTNode {
-                fn from(value: $name) -> Self {
-                    ASTNode::$name(value)
-                }
-            }
-
-            impl From<&$name> for ASTNode {
-                fn from(value: &$name) -> Self {
-                    ASTNode::$name(value.clone())
-                }
-            }
-        )*
-
-        impl ASTNode {
-            pub fn node_type(&self) -> NodeType {
-                match self {
-                    $(ASTNode::$name(_) => NodeType::$name,)*
-                }
-            }
-            pub fn id(&self) -> Option<NodeID> {
-                match self {
-                    $(ASTNode::$name(n) => Some(n.id),)*
-                }
-            }
-        }
-
-        impl Node for ASTNode {
-            fn accept(&self, visitor: &mut impl ASTConstVisitor) -> eyre::Result<()> {
-                match self {
-                    $(ASTNode::$name(n) => n.accept(visitor),)*
-                }
-            }
-            fn accept_metadata(&self, visitor: &mut impl ASTConstVisitor) -> eyre::Result<()> {
-                match self {
-                    $(ASTNode::$name(n) => n.accept_metadata(visitor),)*
-                }
-            }
-            fn accept_id(&self, visitor: &mut impl ASTConstVisitor) -> Result<()> {
-                visitor.visit_node_id(self.id())?;
-                Ok(())
-            }
-        }
-
-        impl ASTNode {
-            pub fn src(&self) -> Option<&str> {
-                match self {
-                    $(ASTNode::$name(node) => Some(&node.src),)*
-                }
-            }
-        }
-
-    };
-}
-
-pub(crate) use generate_ast_methods;
-pub(crate) use generate_capture_methods;
+pub(crate) use generate_capturable_methods;
